feat(icons): move utils out of src
