+++ packages/icons/utils/addAriaLabel.cjs
